fixing: selected samples, ts page rename
