+++ frontend/src/components/individuals/IndividualsDetailContent.tsx
