+++ frontend/src/components/individuals/IndividualsPage.js
@@ -3,7 +3,7 @@ import {Navigate, Route, Routes} from "react-router-dom";
 
 import IndividualEditContent from "./IndividualEditContent";
 import IndividualsListContent from "./IndividualsListContent";
-import IndividualsDetailContent from "./IndividualsDetailContentTS";
+import IndividualsDetailContent from "./IndividualsDetailContent";
 import PageContainer from "../PageContainer";
 
 const IndividualsPage = () => <PageContainer>
